# server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for ChatGameLab - an AI-powered interactive game platform",
        "title": "ChatGameLab API",
        "contact": {
            "name": "ChatGameLab Team",
            "url": "https://chatgamelab.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/institutions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists all institutions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "institutions"
                ],
                "summary": "List institutions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/obj.Institution"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new institution",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "institutions"
                ],
                "summary": "Create institution",
                "parameters": [
                    {
                        "description": "Institution details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.CreateInstitutionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/obj.Institution"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/institutions/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Gets an institution by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "institutions"
                ],
                "summary": "Get institution",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Institution ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/obj.Institution"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Soft-deletes an institution (admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "institutions"
                ],
                "summary": "Delete institution",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Institution ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an institution",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "institutions"
                ],
                "summary": "Update institution",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Institution ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Institution details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.UpdateInstitutionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/obj.Institution"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/institutions/{id}/apikeys": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all API keys shared with an institution (heads and staff only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "institutions"
                ],
                "summary": "Get institution API keys",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Institution ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/obj.ApiKeyShare"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/institutions/{id}/free-use-key": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sets or clears the free-use API key share for an institution.\nAny institution member can use this key to play for free.\nPass null shareId to clear.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "institutions"
                ],
                "summary": "Set institution free-use API key share",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Institution ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Share ID to set (null to clear)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.SetInstitutionFreeUseKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/obj.Institution"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/institutions/{id}/members": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all members of an institution",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "institutions"
                ],
                "summary": "Get institution members",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Institution ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/obj.User"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/institutions/{id}/members/{userID}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes a member from an institution (head or admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "institutions"
                ],
                "summary": "Remove member from institution",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Institution ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/institutions/{id}/prompt-constraints": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sets or clears prompt constraints for an institution (head or admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "institutions"
                ],
                "summary": "Update institution prompt constraints",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Institution ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Prompt constraints",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.UpdateInstitutionPromptConstraintsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/obj.Institution"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httpx.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Machine-readable error code",
                    "type": "string"
                },
                "message": {
                    "description": "Human-readable error message",
                    "type": "string"
                },
                "type": {
                    "description": "Deprecated: use Code instead",
                    "type": "string"
                }
            }
        },
...
        "obj.ApiKeyShare": {
            "type": "object",
            "properties": {
                "apiKey": {
                    "$ref": "#/definitions/obj.ApiKey"
                },
                "apiKeyId": {
                    "type": "string"
                },
                "game": {
                    "$ref": "#/definitions/obj.Game"
                },
                "gameShareId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "institution": {
                    "$ref": "#/definitions/obj.Institution"
                },
                "isPrivateShare": {
                    "type": "boolean"
                },
                "isUserDefault": {
                    "type": "boolean"
                },
                "meta": {
                    "$ref": "#/definitions/obj.Meta"
                },
                "remaining": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/obj.User"
                },
                "workshop": {
                    "$ref": "#/definitions/obj.Workshop"
                }
            }
        },
...
        "obj.Institution": {
            "type": "object",
            "properties": {
                "freeUseAiQualityTier": {
                    "description": "high/medium/low, nil = server default",
                    "type": "string"
                },
                "freeUseApiKeyShareId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/obj.InstitutionMember"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/obj.Meta"
                },
                "name": {
                    "type": "string"
                },
                "promptConstraints": {
                    "type": "string"
                }
            }
        },
        "obj.InstitutionMember": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/obj.Role"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "obj.Role": {
            "type": "string",
            "enum": [
                "admin",
                "head",
                "staff",
                "participant",
                "individual"
            ],
            "x-enum-varnames": [
                "RoleAdmin",
                "RoleHead",
                "RoleStaff",
                "RoleParticipant",
                "RoleIndividual"
            ]
        },
...
        "routes.CreateInstitutionRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "routes.SetInstitutionFreeUseKeyRequest": {
            "type": "object",
            "properties": {
                "shareId": {
                    "type": "string"
                }
            }
        },
...
        "routes.UpdateInstitutionPromptConstraintsRequest": {
            "type": "object",
            "properties": {
                "promptConstraints": {
                    "type": "string"
                }
            }
        },
        "routes.UpdateInstitutionRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token in format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}